refactor base handler
