+++ ads/aqua/extension/base_handler.py
@@ -137,10 +137,3 @@ class AquaAPIhandler(APIHandler):
             return messages[status_code]
         else:
             return default_msg
-
-
-# todo: remove after error handler is implemented
-class Errors(str):
-    INVALID_INPUT_DATA_FORMAT = "Invalid format of input data."
-    NO_INPUT_DATA = "No input data provided."
-    MISSING_REQUIRED_PARAMETER = "Missing required parameter: '{}'"
